🎨 Add Beverages Class
Beverage 배열을 가지고 있는 Beverages 클래스를 만들었다.
